+++ cypress/support/commands.js
@@ -1,3 +1,2 @@
-import "@testing-library/cypress/add-commands";
-
 import "./login";
+import "./visitAndReloadOnceInvalidated";