+++ cypress/support/index.js
@@ -1,1 +1,2 @@
+import "@testing-library/cypress/add-commands";
 import "./commands";