+++ packages/app-page-builder-theme/src/components/Navigation.tsx
@@ -1,14 +1,6 @@
 import React from "react";
 import { Link } from "@webiny/react-router";
 
-function getLink({ url, title }) {
-    if (url.startsWith("/")) {
-        return <Link to={url}>{title}</Link>;
-    }
-
-    return <a href={url}>{title}</a>;
-}
-
 const Navigation = ({ data }) => {
     if (!data) {
         return null;
@@ -23,14 +15,14 @@ const Navigation = ({ data }) => {
                             {item.title}
                             <ul>
                                 {item.children.map((item, index) => {
-                                    return <li key={item.id + index}>{getLink(item)}</li>;
+                                    return <li key={item.id + index}><Link to={item.url}>{item.title}</Link></li>;
                                 })}
                             </ul>
                         </li>
                     );
                 }
 
-                return <li key={item.id + index}>{getLink(item)}</li>;
+                return <li key={item.id + index}><Link to={item.url}>{item.title}</Link></li>;
             })}
         </ul>
     );