+++ packages/app-security/src/admin/components/index.ts
@@ -1,0 +1,1 @@
+export { default as ScopesMultiAutoComplete } from "./ScopesMultiAutoComplete";