+++ packages/app-security/src/admin/plugins/index.ts
@@ -4,6 +4,7 @@ import routes from "./routes";
 import menus from "./menus";
 import secureRouteError from "./secureRouteError";
 import installation from "./installation";
+import scopesList from "./scopesList";
 
 export default [
     // Layout plugins
@@ -11,6 +12,7 @@ export default [
     globalSearchUsers,
     routes,
     menus,
+    scopesList,
     secureRouteError,
     installation
 ];